Fix failing compilation due to duplicate symbols.
Refactor some of the helper functions into their own source file, which should fix compilation failures under GCC. Not sure why it's ok with Clang, but seems like a bad idea all the way around.
