+++ src/components/CMakeLists.txt
@@ -15,7 +15,7 @@ add_library(mcpp-components
         src/CBGShapefileWrapper.cpp include/components/CBGShapefileWrapper.hpp
         src/TemporalMatricies.cpp include/components/TemporalMatricies.hpp
         src/TileWriter.cpp include/components/TileWriter.hpp src/RowProcessor.cpp include/components/RowProcessor.hpp src/detail/CBGOffsetCalculator.cpp include/components/detail/CBGOffsetCalculator.hpp
-        src/VisitMatrixWriter.cpp include/components/VisitMatrixWriter.hpp TemporalGraphs.cpp include/components/TemporalGraphs.hpp include/components/BaseSafegraphMapper.hpp src/BaseSafegraphMapper.cpp include/components/OffsetCalculator.hpp src/detail/CountyOffsetCalculator.cpp include/components/detail/CountyOffsetCalculator.hpp include/components/detail/offset_shared.hpp src/server/CountyShapefileServer.cpp include/components/server/CountyShapefileServer.hpp src/CountyShapefileWrapper.cpp include/components/CountyShapefileWrapper.hpp)
+        src/VisitMatrixWriter.cpp include/components/VisitMatrixWriter.hpp TemporalGraphs.cpp include/components/TemporalGraphs.hpp include/components/BaseSafegraphMapper.hpp src/BaseSafegraphMapper.cpp include/components/OffsetCalculator.hpp src/detail/CountyOffsetCalculator.cpp include/components/detail/CountyOffsetCalculator.hpp include/components/detail/offset_shared.hpp src/server/CountyShapefileServer.cpp include/components/server/CountyShapefileServer.hpp src/CountyShapefileWrapper.cpp include/components/CountyShapefileWrapper.hpp src/detail/helpers.cpp include/components/detail/helpers.hpp)
 
 add_library(MCPP::components ALIAS mcpp-components)
