+++ src/epi-rank/epi-rank.cpp
@@ -97,7 +97,7 @@ int hpx_main(hpx::program_options::variables_map &vm) {
     // Now, compute
     vector<hpx::future<void>> compute_results = engine.compute();
     hpx::wait_all(compute_results);
-    spdlog::debug("Computing completed");
+    spdlog::info("All compution completed");
 
     return hpx::finalize();
 };
